Test calculator.js
